Code Restructure To Reduce Disk Usage and Help Readability
- Move from github-submodule structure to use CMake's native
  dependency management facilities. (https://cmake.org/cmake/help/latest/guide/using-dependencies/index.html). Ideally the dependencies are to be installed using native mechanisms like vcpkg on windows, the linux distribution package manager or homebrew for mac. CMake will find the installed versions as a first option but build the dependencies if not available in the Build Directory at Build time. The linux_build code action is an example of this working in linux.

- Restructure The #include statements to indicate the project source,
  this prevents confusion and is easier to read. It is also a consistent
  way to describe the included header and to reason about the include
  structure. This also is to pave the way to converting to modules

- Explicit Code file declaration in CMake as opposed to File Globbing
  which is only run once during build system generation. When creating a
  new file, it is suggested to specify it in the CMake File which will
  enhance tracking and ensure that it is picked up at build time due to
  changes in the CMake Files.

- Changed Bitmap.h to utf-8 from iso-8859

- Changed from MSVC style of pre-compiled headers, just using plain
  header includes for portability. Another benefit is storage savings as
  the pch file was taking up too much space(in my experience)

+++ ZEngine/ZEngine/Windows/Events/MouseEvent.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <Core/CoreEvent.h>
-#include <Inputs/KeyCode.h>
-#include <ZEngineDef.h>
+#include <ZEngine/Core/CoreEvent.h>
+#include <ZEngine/Windows/Inputs/KeyCode.h>
+#include <ZEngine/ZEngineDef.h>
 
 namespace ZEngine::Windows::Events
 {
@@ -43,7 +43,7 @@ namespace ZEngine::Windows::Events
 
         virtual std::string ToString() const override
         {
-            return fmt::format("MouseButtonPressedEvent : {0}", m_button);
+            return fmt::format("MouseButtonPressedEvent : {0}", static_cast<int32_t>(m_button));
         }
     };
 
@@ -66,7 +66,7 @@ namespace ZEngine::Windows::Events
 
         virtual std::string ToString() const override
         {
-            return fmt::format("MouseButtonReleasedEvent : {0}", m_button);
+            return fmt::format("MouseButtonReleasedEvent : {0}", static_cast<int32_t>(m_button));
         }
     };
 
@@ -136,7 +136,7 @@ namespace ZEngine::Windows::Events
 
         virtual std::string ToString() const override
         {
-            return fmt::format("MouseButtonWheelEvent : {0}", m_button);
+            return fmt::format("MouseButtonWheelEvent : {0}", static_cast<int32_t>(m_button));
         }
 
     private:
